break up middleware tests
